alembic: migrations cleanup
* INSPIR-2370
