+++ .gitignore
@@ -26,7 +26,7 @@ var/
 .installed.cfg
 *.egg
 .python-version
-
+.venv
 # PyInstaller
 #  Usually these files are written by a python script from a template
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
