+++ run-e2e.sh
@@ -1,8 +1,8 @@
 #!/bin/bash -xe
 rc=0
-alias compose="docker-compose -f ./docker-compose.core.yml -f ./e2e/docker-compose.e2e.yml"
-compose up -d --force-recreate
-compose run --rm web ./scripts/setup
-compose stop
-compose up --abort-on-container-exit --exit-code-from e2e || rc=$?
+docker-compose -f ./docker-compose.core.yml up -d --force-recreate
+docker-compose -f ./docker-compose.core.yml exec web ./scripts/setup
+docker-compose -f ./docker-compose.core.yml exec web-next inspirehep db create
+docker-compose -f ./docker-compose.core.yml exec web poetry run inspirehep importer records -d data/records/authors
+docker-compose -f ./docker-compose.core.yml -f ./e2e/docker-compose.e2e.yml run --rm e2e bash -c "cd /opt/e2e && yarn && yarn test" || rc=$?
 exit $rc
